requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx[http2]==0.27.0
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
tenacity==8.2.3
